fix: only owners or operators can exercise and redeem STRAT Options
So long as a STRAT option holder has granted approval to either
the StratOptionExercise or StratOptionRedeemUSDNotional contract,
any third party may either exercise or redeem their option so long
as all other necessary (but caller-independent) conditions are met.

Since all value (either STRAT or ETH) is sent to the option holder,
not the caller, this doesn’t constitute a theft vector. However, it
violates a basic expectation of how token ownership functions,
and could lead to unwanted and badly-timed tax obligations for the
option holder.